backtesting/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
supabase==2.9.1
python-dotenv==1.0.1
pandas==2.2.3
numpy==2.0.2
tenacity==9.0.0
anthropic==0.42.0
